# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BetRoyal Betting & Casino Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /games:
    get:
      operationId: getGames
      tags: [games]
      summary: Get all games
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: provider
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameList"
...
  /games/{id}:
    get:
      operationId: getGame
      tags: [games]
      summary: Get game details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
...
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: Get all users (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
  /admin/users/{id}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get user details (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update user status/details (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/users/{id}/wallet:
    post:
      operationId: adminAdjustWallet
      tags: [admin]
      summary: Manually adjust user wallet (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletAdjustInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
...
  /admin/games/{id}:
    patch:
      operationId: updateAdminGame
      tags: [admin]
      summary: Update game settings (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminGameUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /admin/bonuses:
    get:
      operationId: getAdminBonuses
      tags: [admin]
      summary: Get all bonuses (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bonus"
    post:
      operationId: createAdminBonus
      tags: [admin]
      summary: Create a bonus (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BonusInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bonus"
  /admin/transactions:
    get:
      operationId: getAdminTransactions
      tags: [admin]
      summary: Get all transactions (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
  /admin/reports/daily:
    get:
      operationId: getDailyReport
      tags: [admin]
      summary: Get daily report
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
components:
  schemas:
...
    RefreshTokenInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # USER
    User:
      type: object
      required: [id, username, email, role, status, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin, agent]
        status:
          type: string
          enum: [active, suspended, banned]
        avatar:
          type: ["string", "null"]
        kycStatus:
          type: string
          enum: [none, pending, approved, rejected]
        vipLevel:
          type: integer
        referralCode:
          type: string
        createdAt:
          type: string
...
    Wallet:
      type: object
      required: [id, userId, balance, bonusBalance, currency]
      properties:
        id:
          type: integer
        userId:
          type: integer
        balance:
          type: number
        bonusBalance:
          type: number
        currency:
          type: string
...
    Transaction:
      type: object
      required: [id, userId, type, amount, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [deposit, withdrawal, bet, win, bonus, referral]
        amount:
          type: number
        status:
          type: string
          enum: [pending, completed, failed, cancelled]
        method:
          type: ["string", "null"]
        reference:
          type: ["string", "null"]
        createdAt:
          type: string
    TransactionList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # GAMES
    Game:
      type: object
      required: [id, name, slug, category, provider, rtp, isActive, isFeatured]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        category:
          type: string
          enum: [slots, live_casino, sports, crash, mines, dice, plinko, limbo, keno, coin_flip, lucky_wheel, fishing, cards, blockchain]
        provider:
          type: string
        thumbnail:
          type: ["string", "null"]
        rtp:
          type: number
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        isPopular:
          type: boolean
        maxWin:
          type: ["number", "null"]
        minBet:
          type: number
        maxBet:
          type: number
        isFavorited:
          type: boolean
        playCount:
          type: integer
    GameList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Game"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    Bonus:
      type: object
      required: [id, name, type, value, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [welcome, first_deposit, cashback, daily, vip, referral]
        value:
          type: number
        valueType:
          type: string
          enum: [percentage, fixed]
        minDeposit:
          type: ["number", "null"]
        maxBonus:
          type: ["number", "null"]
        wagerRequirement:
          type: number
        expiresIn:
          type: ["integer", "null"]
        isActive:
          type: boolean
...
    BonusInput:
      type: object
      required: [name, type, value, valueType, wagerRequirement]
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
        value:
          type: number
        valueType:
          type: string
        minDeposit:
          type: number
        maxBonus:
          type: number
        wagerRequirement:
          type: number
        expiresIn:
          type: integer
        isActive:
          type: boolean

    # REFERRALS
...
    AdminUser:
      type: object
      required: [id, username, email, role, status, balance, totalDeposited, totalBets, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
        status:
          type: string
        balance:
          type: number
        bonusBalance:
          type: number
        totalDeposited:
          type: number
        totalWithdrawn:
          type: number
        totalBets:
          type: integer
        kycStatus:
          type: string
        vipLevel:
          type: integer
        createdAt:
          type: string
    AdminUserList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AdminUserUpdate:
      type: object
      properties:
        status:
          type: string
        role:
          type: string
        vipLevel:
          type: integer
    WalletAdjustInput:
      type: object
      required: [amount, type, reason]
      properties:
        amount:
          type: number
        type:
          type: string
          enum: [credit, debit]
        reason:
          type: string
    AdminGameUpdate:
      type: object
      properties:
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        isPopular:
          type: boolean
        rtp:
          type: number
        minBet:
          type: number
        maxBet:
          type: number
    DailyReport:
      type: object
      required: [date, newUsers, deposits, withdrawals, bets, revenue, netRevenue]
      properties:
        date:
          type: string
        newUsers:
          type: integer
        deposits:
          type: number
        withdrawals:
          type: number
        bets:
          type: integer
        totalWagered:
          type: number
        revenue:
          type: number
        netRevenue:
          type: number