# lib/api-zod/src/generated/types/dailyReport.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BetRoyal Betting & Casino Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyReport {
  date: string;
  newUsers: number;
  deposits: number;
  withdrawals: number;
  bets: number;
  totalWagered?: number;
  revenue: number;
  netRevenue: number;
}