# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BetRoyal Betting & Casino Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { GameCategoryProperty } from './gameCategoryProperty';

export interface Game {
  id: number;
  name: string;
  slug: string;
  category: GameCategoryProperty;
  provider: string;
  /** @nullable */
  thumbnail?: string | null;
  rtp: number;
  isActive: boolean;
  isFeatured: boolean;
  isPopular?: boolean;
  /** @nullable */
  maxWin?: number | null;
  minBet?: number;
  maxBet?: number;
  isFavorited?: boolean;
  playCount?: number;
}