# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BetRoyal Betting & Casino Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserKycStatus } from './userKycStatus';
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  username: string;
  email: string;
  role: UserRole;
  status: UserStatus;
  /** @nullable */
  avatar?: string | null;
  kycStatus?: UserKycStatus;
  vipLevel?: number;
  referralCode?: string;
  createdAt: string;
}